Added Final Cart III hack
The FC3 module muxes both controller
ports to the POT lines of the SID.
It is incompatible with two mice
and the way, the Yaumataca simulates
a C1351. This hack enforces to "let go"
of these lines which indirectly might
cause issues with C64 games that support
additional fire buttons.